Set default compression to the fastest version
