+++ src/P434/P434_compressed_api.h
@@ -1,7 +1,7 @@
 /********************************************************************************************
 * SIDH: an efficient supersingular isogeny cryptography library
 *
-* Abstract: API header file for P434 using compression
+* Abstract: API header file for P434 using compression 
 *********************************************************************************************/  
 
 #ifndef P434_COMPRESSED_API_H
@@ -10,33 +10,33 @@
 
 /*********************** Key encapsulation mechanism API ***********************/
 
-#define CRYPTO_SECRETKEYBYTES     240      // MSG_BYTES + SECRETKEY_A_BYTES + CRYPTO_PUBLICKEYBYTES bytes
-#define CRYPTO_PUBLICKEYBYTES     197      // 3*ORDER_B_ENCODED_BYTES + FP2_ENCODED_BYTES + 3 bytes for shared elligator
+#define CRYPTO_SECRETKEYBYTES     350      // MSG_BYTES + SECRETKEY_A_BYTES + CRYPTO_PUBLICKEYBYTES + FP2_ENCODED_BYTES bytes
+#define CRYPTO_PUBLICKEYBYTES     197      // 3*ORDER_B_ENCODED_BYTES + FP2_ENCODED_BYTES + 3 bytes for shared elligator 
 #define CRYPTO_BYTES               16
-#define CRYPTO_CIPHERTEXTBYTES    209     // COMPRESSED_CHUNK_CT + MSG_BYTES bytes
+#define CRYPTO_CIPHERTEXTBYTES    236     // PARTIALLY_COMPRESSED_CHUNK_CT + MSG_BYTES bytes
 
 // Algorithm name
 #define CRYPTO_ALGNAME "SIKEp434_compressed"  
 
 // SIKE's key generation
 // It produces a private key sk and computes the public key pk.
-// Outputs: secret key sk (CRYPTO_SECRETKEYBYTES = 240 bytes)
+// Outputs: secret key sk (CRYPTO_SECRETKEYBYTES = 350 bytes)
 //          public key pk (CRYPTO_PUBLICKEYBYTES = 197 bytes) 
 int crypto_kem_keypair_SIKEp434_compressed(unsigned char *pk, unsigned char *sk);
 
 // SIKE's encapsulation
 // Input:   public key pk         (CRYPTO_PUBLICKEYBYTES = 197 bytes)
 // Outputs: shared secret ss      (CRYPTO_BYTES = 16 bytes)
-//          ciphertext message ct (CRYPTO_CIPHERTEXTBYTES = 209 bytes)
+//          ciphertext message ct (CRYPTO_CIPHERTEXTBYTES = 236 bytes)
 int crypto_kem_enc_SIKEp434_compressed(unsigned char *ct, unsigned char *ss, const unsigned char *pk);
 
 // SIKE's decapsulation
-// Input:   secret key sk         (CRYPTO_SECRETKEYBYTES = 240 bytes)
-//          ciphertext message ct (CRYPTO_CIPHERTEXTBYTES = 209 bytes) 
+// Input:   secret key sk         (CRYPTO_SECRETKEYBYTES = 350 bytes)
+//          ciphertext message ct (CRYPTO_CIPHERTEXTBYTES = 236 bytes) 
 // Outputs: shared secret ss      (CRYPTO_BYTES = 16 bytes)
 int crypto_kem_dec_SIKEp434_compressed(unsigned char *ss, const unsigned char *ct, const unsigned char *sk);
 
-void cryptorun_benchs_SIKEp434_compressed();
+
 
 // Encoding of keys for KEM-based isogeny system "SIKEp434_compressed" (wire format):
 // ---------------------------------------------------------------------------------
@@ -44,10 +44,10 @@ void cryptorun_benchs_SIKEp434_compressed();
 // Elements (a+b*i) over GF(p434^2), where a and b are defined over GF(p434), are encoded as {a, b}, with a in the lowest memory portion.
 //
 // Private keys sk consist of the concatenation of a 16-byte random value, a value in the range [0, 2^216-1] and the public key pk. In the SIKE API, 
-// private keys are encoded in 240 octets in little endian format. 
+// private keys are encoded in 350 octets in little endian format. 
 // Public keys pk consist of 3 values of length OBOB_BITS, one element in GF(p434^2) and 2 bytes. In the SIKE API, pk is encoded in 197 octets. 
 // Ciphertexts ct consist of the concatenation of 3 values of length OALICE_BITS, one element in GF(p434^2), 2 bytes and a 16-byte value. In the SIKE API, 
-// ct is encoded in 3*27 + 110 + 2 + 16 = 209 octets.  
+// ct is encoded in 3*27 + 110 + 2 + 16 = 236 octets.  
 // Shared keys ss consist of a value of 16 octets.
 
 
@@ -73,7 +73,7 @@ void random_mod_order_B_SIDHp434(unsigned char* random_digits);
 // Alice's ephemeral public key generation
 // Input:  a private key PrivateKeyA in the range [0, 2^216 - 1], stored in 27 bytes. 
 // Output: the public key PublicKeyA consisting of 3 values of length OALICE_BITS, one element in GF(p434^2) and 2 bytes encoded in 197 bytes.
-int EphemeralKeyGeneration_A_SIDHp434_Compressed(const unsigned char* PrivateKeyA, unsigned char* PublicKeyA);
+int EphemeralKeyGeneration_A_SIDHp434_Compressed(unsigned char* PrivateKeyA, unsigned char* PublicKeyA);
 
 // Bob's ephemeral key-pair generation
 // It produces a private key PrivateKeyB and computes the public key PublicKeyB.
@@ -86,7 +86,7 @@ int EphemeralKeyGeneration_B_SIDHp434_Compressed(const unsigned char* PrivateKey
 // Inputs: Alice's PrivateKeyA is an integer in the range [0, 2^216 - 1], stored in 27 bytes. 
 //         Bob's PublicKeyB consists of 3 values of length OALICE_BITS, one element in GF(p434^2) and 2 bytes encoded in 197 bytes.
 // Output: a shared secret SharedSecretA that consists of one element in GF(p434^2) encoded in 110 bytes.
-int EphemeralSecretAgreement_A_SIDHp434_Compressed(const unsigned char* PrivateKeyA, const unsigned char* PublicKeyB, unsigned char* SharedSecretA);
+int EphemeralSecretAgreement_A_SIDHp434_Compressed(const unsigned char* PrivateKeyA, const unsigned char* PublicKeyB, unsigned char* SharedSecretA, unsigned char* phiBKA_t);
 
 // Bob's ephemeral shared secret computation
 // It produces a shared secret key SharedSecretB using his secret key PrivateKeyB and Alice's public key PublicKeyA
