+++ src/P610/P610_compressed.c
@@ -6,6 +6,7 @@
 
 #include "P610_compressed_api.h"
 #define COMPRESS
+#define COMPRESS_SPEED
 #include "P610_internal.h"
 
 
@@ -174,13 +175,13 @@ const unsigned int ph2_path[PLEN_2] = { // w_2 = 5
   35, 35, 35, 35, 35, 36, 37, 38, 39, 40, 41, 41, 42
 };
 
-const unsigned int ph3_path[PLEN_3] = { 
+const unsigned int ph3_path[PLEN_3] = {
 #if W_3 == 4
 0, 0, 1, 2, 3, 4, 4, 5, 5, 6, 7, 7, 8, 9, 10, 10, 11, 12, 13, 14, 14, 14, 15, 16, 17, 18, 19, 19, 19, 19, 20, 21, 22, 23, 24, 25, 26, 26, 26, 26, 26, 27, 28, 29, 30, 31, 32, 33, 34  
 #elif W_3 == 6  
     0, 0, 1, 2, 3, 4, 5, 6, 6, 7, 7, 8, 9, 9, 10, 11, 12, 12, 13, 14, 15, 16, 16, 17, 18, 19, 
     20, 21, 21, 22, 23, 24, 25
-#endif    
+#endif   
 };
 
 // Entangled bases related static tables and parameters
@@ -366,11 +367,11 @@ const uint64_t v_3_torsion[20][2 * NWORDS64_FIELD] =
 #define EphemeralKeyGeneration_A      EphemeralKeyGeneration_A_SIDHp610_Compressed
 #define EphemeralKeyGeneration_B      EphemeralKeyGeneration_B_SIDHp610_Compressed
 #define EphemeralSecretAgreement_A    EphemeralSecretAgreement_A_SIDHp610_Compressed
-#define EphemeralSecretAgreement_B    EphemeralSecretAgreement_B_SIDHp610_Compressed 
+#define EphemeralSecretAgreement_B    EphemeralSecretAgreement_B_SIDHp610_Compressed
 #define crypto_kem_keypair            crypto_kem_keypair_SIKEp610_compressed
 #define crypto_kem_enc                crypto_kem_enc_SIKEp610_compressed
 #define crypto_kem_dec                crypto_kem_dec_SIKEp610_compressed
-#define cryptorun_benchs              cryptorun_benchs_SIKEp610_compressed
+
 
 #include "../fpx.c"
 #include "../ec_isogeny.c"
