+++ src/P751/P751_compressed_api.h
@@ -1,19 +1,19 @@
 /********************************************************************************************
 * SIDH: an efficient supersingular isogeny cryptography library
 *
-* Abstract: API header file for P751 using compression
+* Abstract: API header file for P751 using compression with ciphertext vs decaps. speed tradeoff
 *********************************************************************************************/  
 
-#ifndef P751_COMPRESSED_API_H
-#define P751_COMPRESSED_API_H
+#ifndef P751_COMPRESSED_SPEED_API_H
+#define P751_COMPRESSED_SPEED_API_H
     
 
 /*********************** Key encapsulation mechanism API ***********************/
 
-#define CRYPTO_SECRETKEYBYTES     414      // MSG_BYTES + SECRETKEY_A_BYTES + CRYPTO_PUBLICKEYBYTES bytes
-#define CRYPTO_PUBLICKEYBYTES     335      // 3*ORDER_B_ENCODED_BYTES + FP2_ENCODED_BYTES + 2 bytes for shared elligator
+#define CRYPTO_SECRETKEYBYTES     602      // MSG_BYTES + SECRETKEY_A_BYTES + CRYPTO_PUBLICKEYBYTES + FP2_ENCODED_BYTES bytes
+#define CRYPTO_PUBLICKEYBYTES     335      // 3*ORDER_B_ENCODED_BYTES + FP2_ENCODED_BYTES + 3 bytes for shared elligator
 #define CRYPTO_BYTES               32
-#define CRYPTO_CIPHERTEXTBYTES    363      // COMPRESSED_CHUNK_CT + MSG_BYTES bytes     
+#define CRYPTO_CIPHERTEXTBYTES    410      // COMPRESSED_CHUNK_CT + MSG_BYTES bytes     
 
 // Algorithm name
 #define CRYPTO_ALGNAME "SIKEp751_compressed" 
@@ -27,16 +27,15 @@ int crypto_kem_keypair_SIKEp751_compressed(unsigned char *pk, unsigned char *sk)
 // SIKE's encapsulation
 // Input:   public key pk         (CRYPTO_PUBLICKEYBYTES = 335 bytes)
 // Outputs: shared secret ss      (CRYPTO_BYTES = 32 bytes)
-//          ciphertext message ct (CRYPTO_CIPHERTEXTBYTES = 363 bytes) 
+//          ciphertext message ct (CRYPTO_CIPHERTEXTBYTES = 410 bytes) 
 int crypto_kem_enc_SIKEp751_compressed(unsigned char *ct, unsigned char *ss, const unsigned char *pk);
 
 // SIKE's decapsulation
 // Input:   secret key sk         (CRYPTO_SECRETKEYBYTES = 414 bytes)
-//          ciphertext message ct (CRYPTO_CIPHERTEXTBYTES = 363 bytes) 
+//          ciphertext message ct (CRYPTO_CIPHERTEXTBYTES = 410 bytes) 
 // Outputs: shared secret ss      (CRYPTO_BYTES = 32 bytes)
 int crypto_kem_dec_SIKEp751_compressed(unsigned char *ss, const unsigned char *ct, const unsigned char *sk);
 
-void cryptorun_benchs_SIKEp610_compressed();
 
 
 // Encoding of keys for KEM-based isogeny system "SIKEp751" (wire format):
@@ -47,8 +46,8 @@ void cryptorun_benchs_SIKEp610_compressed();
 // Private keys sk consist of the concatenation of a 32-byte random value, a value in the range [0, 2^372-1] and the public key pk. In the SIKE API, 
 // private keys are encoded in 414 octets in little endian format.  
 // Public keys pk consist of 3 values of length OBOB_BITS, one element in GF(p751^2) and 2 bytes. In the SIKE API, pk is encoded in 335 octets.
-// Ciphertexts ct consist of the concatenation of 3 values of length OALICE_BITS, one element in GF(p751^2), 2 bytes and a 32-byte value. In the SIKE API, 
-// ct is encoded in 3*47 + 188 + 2 + 32 = 363 octets.  
+// Ciphertexts ct consist of the concatenation of 4 values of length OALICE_BITS, one element in GF(p751^2), 2 bytes and a 32-byte value. In the SIKE API, 
+// ct is encoded in 3*47 + 188 + 2 + 32 = 410 octets.  
 // Shared keys ss consist of a value of 32 octets.
 
 
@@ -74,7 +73,7 @@ void random_mod_order_B_SIDHp751(unsigned char* random_digits);
 // Alice's ephemeral public key generation
 // Input:  a private key PrivateKeyA in the range [0, 2^372 - 1], stored in 47 bytes. 
 // Output: the public key PublicKeyA consisting of 3 GF(p751^2) elements encoded in 335 bytes.
-int EphemeralKeyGeneration_A_SIDHp751_Compressed(const unsigned char* PrivateKeyA, unsigned char* PublicKeyA);
+int EphemeralKeyGeneration_A_SIDHp751_Compressed(unsigned char* PrivateKeyA, unsigned char* PublicKeyA);
 
 // Bob's ephemeral key-pair generation
 // It produces a private key PrivateKeyB and computes the public key PublicKeyB.
@@ -87,7 +86,7 @@ int EphemeralKeyGeneration_B_SIDHp751_Compressed(const unsigned char* PrivateKey
 // Inputs: Alice's PrivateKeyA is an integer in the range [0, 2^372 - 1], stored in 47 bytes. 
 //         Bob's PublicKeyB consists of 3 GF(p751^2) elements encoded in 335 bytes.
 // Output: a shared secret SharedSecretA that consists of one element in GF(p751^2) encoded in 188 bytes.
-int EphemeralSecretAgreement_A_SIDHp751_Compressed(const unsigned char* PrivateKeyA, const unsigned char* PublicKeyB, unsigned char* SharedSecretA);
+int EphemeralSecretAgreement_A_SIDHp751_Compressed(const unsigned char* PrivateKeyA, const unsigned char* PublicKeyB, unsigned char* SharedSecretA, unsigned char* phiBKA_t);
 
 // Bob's ephemeral shared secret computation
 // It produces a shared secret key SharedSecretB using his secret key PrivateKeyB and Alice's public key PublicKeyA
